refactor: split bayesian_var.py (892L) into subpackage (#223)
- bayesian_var/__init__.py (29L) — re-exports all public + tested internal names
- bayesian_var/priors.py (138L) — MinnesotaPrior, NormalWishartPrior, prior helpers
- bayesian_var/results.py (39L) — BayesianVARResult dataclass
- bayesian_var/inference.py (164L) — _analytical_posterior, _gibbs_sample
- bayesian_var/model.py (403L) — BayesianVAR class, _build_var_matrices, bayesian_var()
- All files under 500 lines (max 403L vs original 892L)
- All existing import paths preserved via __init__.py re-exports
- 7 new structural tests, 67/67 total tests pass
